prototype kinetic osd backend

+++ src/os/FileJournal.cc
@@ -1758,7 +1758,12 @@ FileJournal::read_entry_result FileJournal::do_read_entry(
   // ok!
   if (seq)
     *seq = h->seq;
-  journalq.push_back(pair<uint64_t,off64_t>(h->seq, pos));
+
+  // works around an apparent GCC 4.8(?) compiler bug about unaligned
+  // bind by reference to (packed) h->seq
+  journalq.push_back(
+    pair<uint64_t,off64_t>(static_cast<uint64_t>(h->seq),
+			   static_cast<off64_t>(pos)));
 
   if (next_pos)
     *next_pos = pos;
